+++ gen/aws/generators/paramsdoc.go
@@ -142,6 +142,7 @@ func loadAllRefs() map[string]string {
 		filepath.Join("cloudformation", "2010-05-15", "docs-2.json"),
 		filepath.Join("ecr", "2015-09-21", "docs-2.json"),
 		filepath.Join("ecs", "2014-11-13", "docs-2.json"),
+		filepath.Join("application-autoscaling", "2016-02-06", "docs-2.json"),
 	}
 
 	entriesC := make(chan *entries)