+++ template/revert.go
@@ -102,6 +102,10 @@ func (te *Template) Revert() (*Template, error) {
 				case "accesskey":
 					params = append(params, fmt.Sprintf("id=%s", quoteParamIfNeeded(cmd.CmdResult)))
 					params = append(params, fmt.Sprintf("user=%s", quoteParamIfNeeded(cmd.Params["user"])))
+				case "appscalingtarget":
+					params = append(params, fmt.Sprintf("dimension=%s", quoteParamIfNeeded(cmd.Params["dimension"])))
+					params = append(params, fmt.Sprintf("resource=%s", quoteParamIfNeeded(cmd.Params["resource"])))
+					params = append(params, fmt.Sprintf("service-namespace=%s", quoteParamIfNeeded(cmd.Params["service-namespace"])))
 				case "loginprofile":
 					params = append(params, fmt.Sprintf("username=%s", quoteParamIfNeeded(cmd.Params["username"])))
 				case "container":
@@ -221,6 +225,10 @@ func isRevertible(cmd *ast.CommandNode) bool {
 	}
 
 	if cmd.Entity == "container" && cmd.Action == "create" {
+		return true
+	}
+
+	if cmd.Entity == "appscalingtarget" && cmd.Action == "create" {
 		return true
 	}